windows: store app resources and data struct in Rc<AppState>
Individual Windows will need to store a reference back to the AppState
in their user data so it is accessible from the WNDPROC, so store it in
an Rc.

Store the data struct in a RefCell so that it can be accessed mutably
during event handlers.

Finally, store the data struct as an Option<T>, so that it can be taken
and dropped in App::drop(), which will break reference cycles due to
Windows stored in T.