windows: close window in Window::drop()
Call DestroyWindow() in WindowInner::drop(), and then free the
WindowState in the WM_DESTROY event.

One subtlety is that when DestroyWindow() is called, Windows
synchronously deliverse the WM_DESTROY event during that call, so if
this happens during an event handler, we can't free the WindowState yet.
We solve this by storing the WindowState in an Rc and cloning it for the
duration of the WNDPROC.