windows: store handler in Box<dyn HandleEvent>
Currently, WindowState<T> is generic over the app data T, but Window is
type-erased. This makes it imposslbe to access the WindowState directly
from the Window struct. Introduce a HandleEvent trait and move
WindowState's Rc<AppState<T>> reference into a Handler<T, H> struct
alongside the event handler callback, and store this as a Box<dyn
HandleEvent> in the WindowState.

This makes WindowState fully type-erased and thus accessible via the
Window struct, which also makes it possible to move all Window cleanup
out of the WM_DESTROY event and into WindowInner::drop(). The situation
regarding reference counting remains the same (the window's user data
pointer holds one reference to the Rc<WindowState> which is freed in
WindowInner::drop(), and the WNDPROC also retains a reference for the
duration of the function).