x11: window creation
Basic window creation. Store window ID and a reference to the AppState
struct in WindowInner. Store window handler closures in a HashMap in
AppState.

Because Window doesn't have a type parameter for the app data type T,
but handlers need to remain typed so that they can be passed &mut T when
handling events, the handlers HashMap is stored in AppState as a trait
object, so that App<T> can hold a typed reference to it and Window can
hold a type-erased one.