windows: stop using BeginPaint/EndPaint in WM_PAINT
Since we are switching to an API where exposed rectangles are explicitly
passed in the Expose event, and dirty rectangles are explicitly passed
to the update_contents function, we don't want to use
BeginPaint/EndPaint, since they automatically set up a clip region for
us. Instead, we just directly send the Expose event and then delegate to
DefWindowProc.