windows: store Rc<WindowState> in WindowInner
Previously the WindowInner held only the HWND and had to call
GetWindowLongPtr to access the WindowState. Instead, store a direct
reference to the WindowState in WindowInner, and store the HWND in the
WindowState.

Also, instead of storing a Box<dyn HandleEvent> in the WindowState, make
the WindowState be a DST and store the Handler struct inline (similarly
to the X11 backend). Since this makes a type-erased reference to the
WindowState require a wide pointer (*mut WindowState<dyn HandleEvent>),
change the window's user data to a boxed wide pointer
(Box<Rc<WindowState<dyn HandleEvent>>>).