switch to 2021 edition; enable winapi std feature
The 2021 edition switches Cargo's feature resolver to version 2, which
means crate feature sets are no longer unified across separate targets,
dev-dependencies, build-dependencies, proc macros, and standard
dependencies. The previous resolver behavior was indirectly enabling
winapi's std feature, which makes it use the same c_void type as std.

Portlight was implicitly relying on this behavior, and since Cargo's
resolver version is a global option, Portlight fails to build as a
dependency of a crate which is using resolver version 2 (e.g. because it
is on edition 2021). This change fixes the problem by explicitly
enabling the "std" feature, as well as changing the edition to 2021 so
that similar problems will be caught in the future.