x11: remove generic T parameter from AppState and WindowState
Refactor X11 backend to make AppState and WindowState no longer be
generic over the app data type. Store handlers as boxed closures which
take a `&mut dyn Any` for the app data and then downcast it to the
appropriate T. Finally, have WindowInner hold a `Weak<AppState>` and a
window ID, close all windows when the `App` is dropped, and panic in
methods on `Window` if they are called after the `App` has been dropped.