win32: fix issue with first expose event not being handled
If a window is created and shown during the App::new builder callback,
the resulting expose event is not able to be handled yet, but we were
still validating the dirty region (by falling through to the
DefWindowProc. If we only validate the dirty region when the expose
event is successfully handled, WM_PAINT messages received during the
builder callback will be sent again later when they can be handled.