cocoa: store app data as a Box<dyn Any>
Because Window does not have a type parameter for the app data, it needs to be
type-erased in some way. One possibility would be for WindowInner to hold a
Weak<AppState<dyn Any>>, but unfortunately this requires T: ?Sized, and T is
stored in an Option which has a T: Sized bound. Just storing the app data as a
Box<dyn Any> and downcasting it where necessary is the most straightforward
solution for now.