x11: Windows hold Rc<WindowState> instead of Weak<AppState>
Instead of having Windows hold a Weak<AppState> and closing all windows
when the App is dropped, have Windows hold an Rc<WindowState> directly,
which in turn holds an Rc<AppState>. This means that if a Window
outlives the App, the AppState will just get cleaned up when the last
Window is dropped.

When the Window struct is dropped, it still removes its corresponding
Rc<WindowState> from the set of windows in AppState, breaking the
reference cycle between AppState and WindowState.