make position optional in WindowOptions
Make the position field in WindowOptions into an Option, and if it is
unset, position the newly created window in a default location. On
Windows, this means passing CW_USEDEFAULT, and on macOS, this means
calling [NSWindow center]. X11 behavior seems to be dependent on the
window manager and could use some more investigation.