remove app data. all state is now stored in window and timer handlers
Remove the app data (T in App<T>). Building the App struct no longer
requires a builder closure that constructs the app data. App now has a
context method for getting access to the AppContext outside an event
handler, which makes it possible to open windows before entering the
event loop. Window handlers are also now passed a &Window, which makes
it possible to call methods on the window without passing it in via an
Rc or similar.

These APIs will get cleaned up in the future. AppContext will become an
AppHandle which implements Clone, and window and timer handlers will be
passed a WindowContext and TimerContext which allow getting access to
both the AppHandle and the Window and Timer objects respectively.