replace AppContext with WindowContext, TimerContext, and AppHandle
Introduce an AppHandle struct which implements Clone (unlike App) but
does not keep the app alive (i.e. if App is dropped, all windows are
closed regardless of whether AppHandles still exist).

Also introduce WindowContext and TimerContext structs which are passed to
window and timer handlers respectively and which provide access to an
AppHandle and the relevant Window or Timer struct.