win32: simplify WNDPROC catch_unwind logic
Instead of wrapping each individual event dispatch in a call to
AppState::catch_unwind, wrap the whole body of the WNDPROC in
std::panic::catch_unwind and then use AppState::propagate_panic if a
panic occurred. This reduces the number of WindowState::from_raw calls
and simplifies the overall control flow of the WNDPROC.

Also, use catch_unwind when dropping the Rc<WindowState> or Rc<AppState>
at the end of the WNDPROC. Previously, if a panic occurred when dropping
the Rc, it would have unwound across the FFI boundary.